refactor: pass data using props
