+++ app/Http/Controllers/Sliders/SliderController.php
@@ -1,0 +1,98 @@
+<?php
+
+namespace App\Http\Controllers\Sliders;
+
+use App\Http\Controllers\Controller;
+use Inertia\Inertia;
+
+class SliderController extends Controller
+{
+    public function index()
+    {
+        $slides = [
+            [
+                'image' => 'images/slider-1/1.png',
+                'title' => 'Бесплатно!',
+                'text' => '1 активная акция всегда бесплатно! Нужно больше и сразу? Наши тарифы вас приятно удивят!'
+            ],
+            [
+                'image' => 'images/slider-1/2.png',
+                'title' => 'Оригинально!',
+                'text' => 'Разработанный нами онлайн конструктор акций, позволит вам без труда создать потрясающую и подробную акцию, индивидуальный шаблон для каждой акции, а также задать множество дополнительных параметров (промокоды, бесплатная доставка, способ расчета скидки и многое другое)! Запустить ее в считанные минуты! Мы постарались продумать там все до мелочей!'
+            ],
+            [
+                'image' => 'images/slider-1/3.png',
+                'title' => 'Удобно!',
+                'text' => 'Все акции, которые вы запускали ранее, после завершения будут автоматически сохранены в личном кабинете, а их перезапуск займет пару кликов. Это очень быстро! Архив запущенных и прошедших акции всегда сохранен в Вашем личном кабинете!'
+            ],
+            [
+                'image' => 'images/slider-1/4.png',
+                'title' => 'Автоматизировано!',
+                'text' => 'Благодаря нашим расширенным/платным возможностям Ваши акции будут подниматься в ТОП сами – нет необходимости в ручном редактировании и продлении. Отдыхайте! Пусть ваши акции работают сам'
+            ],
+            [
+                'image' => 'images/slider-1/5.png',
+                'title' => 'Логично!',
+                'text' => 'Вы сами выбираете, в какие дни и какие часы, ваши акции будут показаны на сайте, что очень удобно - показывать акции когда это необходимо и скрывать их, когда в этом нет нужды.'
+            ],
+            [
+                'image' => 'images/slider-1/6.png',
+                'title' => 'Красиво!',
+                'text' => 'Предварительный просмотр перед стартом акции позволит Вам оценить её внешний вид и визуальное оформление и только потом запустить! Это позволит вам сделать акцию максимально красиво и с первой попытки'
+            ],
+            [
+                'image' => 'images/slider-1/7.png',
+                'title' => 'Глобально!',
+                'text' => 'Абсолютно неважно где территориально расположен Ваш бизнес, каков его размер и организационная форма - c GM может начать зарабатывать любой желающий, от обычного фрилансера до огромной компании!'
+            ],
+            [
+                'image' => 'images/slider-1/8.png',
+                'title' => 'Экономно!',
+                'text' => 'Тарифы распространяются на все сервисы GM...'
+            ],
+            [
+                'image' => 'images/slider-1/9.png',
+                'title' => 'Сгенерировано!',
+                'text' => 'Собственно разработанный генератор, позволит вам генерировать тысячи денежных кодов, которые вы можете разыгрывать среди ваших клиентов, подписчиков и участников. А так же выдавать таким образом кэшбэк, со своего счета на счет вашего клиента!'
+            ]
+        ];
+
+        $reviews = [
+            [
+                'text' => 'Давно пользуюсь и все норм. Ни за что не плачу, спокойно продаю что нужно и покупаю что увижу. Замутили неплохую доставку – теперь можно даже при получении оплачивать и не бояться.',
+                'name' => 'Павел Ларионов',
+                'position' => 'Директор сети кофитериев “Лайк”',
+                'avatar' => 'images/card-avatars/1.png'
+            ],
+            [
+                'text' => 'Несколько лет трудимся с ними вместе и очень довольны результатами совместной работы. Хотим отметить ответственность и профессионализм их сотрудников. Сервис, который они разработали, обладает великолепным функционалом и помогает людям находить самое лучшее.',
+                'name' => 'Дмитрий Валентинов',
+                'position' => 'Основатель школы макияжа и стиля, организатор мероприятий в сфере красоты и моды.',
+                'avatar' => 'images/card-avatars/2.png'
+            ],
+            [
+                'text' => 'Из всех торговых площадок, на которых я пробовала продавать, GMсамая массовая и удобная. Тут на объявление реагируют быстро и продажи идут практически сразу после подачи заявки. Все, конечно, от товара зависит. В моем случае – это корейская косметика. Дела пошли сразу, я довольна. На других сайтах ждала клиента неделями, а тут каждый день практически что-то продаю.',
+                'name' => 'Илона Семенова',
+                'position' => 'Основатель школы макияжа и стиля Корейской косметики “КореяШоп”',
+                'avatar' => 'images/card-avatars/3.png'
+            ],
+            [
+                'text' => 'Давно пользуюсь и все норм. Ни за что не плачу, спокойно продаю что нужно и покупаю что увижу. Замутили неплохую доставку – теперь можно даже при получении оплачивать и не бояться.',
+                'name' => 'Павел Ларионов',
+                'position' => 'Директор сети кофитериев “Лайк”',
+                'avatar' => 'images/card-avatars/4.png'
+            ],
+            [
+                'text' => 'Несколько лет трудимся с ними вместе и очень довольны результатами совместной работы. Хотим отметить ответственность и профессионализм их сотрудников. Сервис, который они разработали, обладает великолепным функционалом и помогает людям находить самое лучшее.',
+                'name' => 'Дмитрий Валентинов',
+                'position' => 'Основатель школы макияжа и стиля, организатор мероприятий в сфере красоты и моды.',
+                'avatar' => 'images/card-avatars/5.png'
+            ]
+        ];
+
+        return Inertia::render('Welcome', [
+            'slides' => $slides,
+            'reviews' => $reviews,
+        ]);
+    }
+}
