+++ routes/web.php
@@ -2,12 +2,12 @@
 
 declare(strict_types=1);
 
+use App\Http\Controllers\Sliders\SliderController;
 use Illuminate\Support\Facades\Route;
 use Inertia\Inertia;
 
-Route::get('/', fn () => Inertia::render('Welcome'))->name('home');
 
-//Route::get('/lending', fn () => Inertia::render('landing/Index'))->name('lending');
+Route::get('/', [SliderController::class, 'index'])->name('home');
 
 Route::get('dashboard', fn () => Inertia::render('Dashboard'))->middleware(['auth', 'verified'])->name('dashboard');
